# internal/webserver/swaggerdocs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API for the honey web UI (`honey web`). Authenticate with the same token printed at startup, or `HONEY_WEB_TOKEN`. WebSocket endpoints `GET /ws/ssh` and `GET /ws/pve-qemu-vnc` are not described by this OpenAPI document.",
        "title": "Honey Web API",
        "contact": {
            "name": "honey"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8765",
    "basePath": "/",
    "paths": {
...
        "/api/v1/tunnels": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tunnels"
                ],
                "summary": "List tunnels",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/webserver.TunnelsListResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tunnels"
                ],
                "summary": "Start tunnel",
                "parameters": [
                    {
                        "description": "ssh_user, record, mapping (e.g. 8080:localhost:8080)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/webserver.StartTunnelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/webserver.TunnelStartResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/tunnels/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tunnels"
                ],
                "summary": "Stop tunnel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "tunnel id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/webserver.TunnelDeleteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/tunnels/{id}/logs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tunnels"
                ],
                "summary": "Tunnel logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "tunnel id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/webserver.TunnelLogsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "hosts.Record": {
            "type": "object",
            "properties": {
                "extra_ips": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "meta": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "primary_ip": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "zone": {
                    "type": "string"
                }
            }
        },
...
        "webserver.ActiveTunnel": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "host": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "mapping": {
                    "type": "string"
                },
                "record_key": {
                    "type": "string"
                },
                "started_at": {
                    "type": "string"
                }
            }
        },
...
        "webserver.StartTunnelRequest": {
            "type": "object",
            "properties": {
                "mapping": {
                    "type": "string"
                },
                "record": {
                    "$ref": "#/definitions/hosts.Record"
                },
                "ssh_user": {
                    "type": "string"
                }
            }
        },
...
        "webserver.TunnelDeleteResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                }
            }
        },
        "webserver.TunnelLogsResponse": {
            "type": "object",
            "properties": {
                "logs": {
                    "type": "string"
                }
            }
        },
        "webserver.TunnelStartResponse": {
            "type": "object",
            "properties": {
                "tunnel": {
                    "$ref": "#/definitions/webserver.ActiveTunnel"
                }
            }
        },
        "webserver.TunnelsListResponse": {
            "type": "object",
            "properties": {
                "tunnels": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/webserver.ActiveTunnel"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Use `Authorization: Bearer \u003ctoken\u003e` where `\u003ctoken\u003e` is the web UI token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "HoneyTokenHeader": {
            "description": "Alternative header: `X-Honey-Token: \u003ctoken\u003e`.",
            "type": "apiKey",
            "name": "X-Honey-Token",
            "in": "header"
        },
        "TokenQuery": {
            "description": "Optional query token (same value) for URLs: `?token=\u003ctoken\u003e`.",
            "type": "apiKey",
            "name": "token",
            "in": "query"
        }
    }
}